dev/content-indexer/requirements.txt
kafka-python~=2.0.2
minio~=7.1.2
requests~=2.26.0
